chatGPT_api/requirements.txt
Flask
python-dotenv
langchain
openai